Change listContacts() name to showContacts(). this func print a list of existing contatcts' name with their index in contactList[].
Def new listContacts(). this new func initialize contactList[] with keys in contacts{} and make key Capitalize or lowecase according to given capital variable.

Edit menu() according to name change in showContacts().

Edit checkExist() to use new listContacts().

Edit showContacts() to use new listContacts().

Edit addPerson() to Capitalize input name to be compatible with new listContacts(), contactList[] and checkExist().

Edit editPerson() to Capitallize input name and newName and lowercase name in checkExist IF to be compatible with new listContacts(), contactList[] and checkExist().

Edit delPerson() to Capitallize input name and lowercase name in checkExist IF to be compatible with new listContacts(), contactList[] and checkExist().